# api/openapi.yaml
openapi: 3.1.0
info:
  title: OpenJourney API
  version: 0.1.0
  description: Phase 1 event ingestion and profile projection API.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://localhost:8080
paths:
...
  /v1/events/batch:
    post:
      operationId: acceptEventBatch
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [events]
              additionalProperties: false
              properties:
                events:
                  type: array
                  minItems: 1
                  maxItems: 75
                  items:
                    $ref: "#/components/schemas/Event"
      responses:
        "202":
          description: The events are durable and queued for projection.
          content:
            application/json:
              schema:
                type: object
                required: [status, event_ids]
                properties:
                  status:
                    const: accepted
                  event_ids:
                    type: array
                    items:
                      type: string
                      format: uuid
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /v1/auth/login:
    post:
      operationId: loginLocalOperator
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              additionalProperties: false
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 12 }
      responses:
        "200":
          description: Local self-hosted operator session.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "500": { $ref: "#/components/responses/Error" }
  /v1/auth/logout:
    post:
      operationId: logoutLocalOperator
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Local self-hosted operator session revoked. API keys and OIDC tokens are ignored.
        "500": { $ref: "#/components/responses/Error" }
  /v1/profiles/{externalID}:
    get:
      operationId: getProfile
      security:
        - bearerAuth: []
      parameters:
        - name: externalID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Current profile and consent projection.
          content:
            application/json:
              schema:
                type: object
                required: [profile, consents]
                properties:
                  profile:
                    $ref: "#/components/schemas/Profile"
                  consents:
                    type: array
                    items:
                      $ref: "#/components/schemas/Consent"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /v1/schemas:
    get:
      operationId: listEventSchemas
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Registered event schemas.
          content:
            application/json:
              schema:
                type: object
                properties:
                  schemas:
                    type: array
                    items: { $ref: "#/components/schemas/EventSchema" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
    post:
      operationId: createEventSchema
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventSchemaInput" }
      responses:
        "201":
          description: Schema registered.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventSchema" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /v1/api-keys:
    get:
      operationId: listAPIKeys
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: API key metadata; secrets are never returned.
          content:
            application/json:
              schema:
                type: object
                required: [api_keys]
                properties:
                  api_keys:
                    type: array
                    items: { $ref: "#/components/schemas/APIKey" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
    post:
      operationId: createAPIKey
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, scopes]
              additionalProperties: false
              properties:
                name: { type: string }
                scopes: { type: array, minItems: 1, items: { type: string } }
                expires_at: { type: string, format: date-time }
      responses:
        "201":
          description: API key metadata and the one-time secret.
          content:
            application/json:
              schema:
                type: object
                required: [api_key, secret]
                properties:
                  api_key: { $ref: "#/components/schemas/APIKey" }
                  secret: { type: string }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /v1/api-keys/{id}:
    delete:
      operationId: revokeAPIKey
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: API key revoked. }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /v1/privacy/requests:
    post:
      operationId: createPrivacyRequest
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [external_id, request_type]
              additionalProperties: false
              properties:
                external_id: { type: string }
                request_type: { enum: [export, delete] }
      responses:
        "202":
          description: Durable privacy operation accepted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PrivacyRequest" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /v1/privacy/requests/{id}:
    get:
      operationId: getPrivacyRequest
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Privacy operation state.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PrivacyRequest" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /v1/operations/queues:
    get:
      operationId: getQueueStatus
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Tenant queue depth by state. }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /v1/operations/dlq:
    get:
      operationId: listDeadLetters
      security: [{ bearerAuth: [] }]
      parameters:
        - name: queue
          in: query
          schema: { enum: [projection, outbox, operations] }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
      responses:
        "200":
          description: Tenant-scoped dead-letter queue items.
          content:
            application/json:
              schema:
                type: object
                required: [dead_letters]
                properties:
                  dead_letters:
                    type: array
                    items: { $ref: "#/components/schemas/DeadLetterItem" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /v1/operations/dlq/{queue}/{id}/retry:
    post:
      operationId: retryDeadLetter
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: queue, in: path, required: true, schema: { enum: [projection, outbox, operations] } }
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Dead-letter item was moved back to pending. }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /v1/operations/dlq/{queue}/{id}/discard:
    post:
      operationId: discardDeadLetter
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: queue, in: path, required: true, schema: { enum: [projection, outbox, operations] } }
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Dead-letter item was discarded. }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /v1/operations/replay/verify:
    post:
      operationId: verifyProfileReplay
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Live and replayed projections match.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReplayReport" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409":
          description: Projection drift detected.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReplayReport" }
  /v1/roles:
    get:
      operationId: listRoles
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Tenant roles. }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
    post:
      operationId: createRole
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, permissions]
              properties:
                name: { type: string }
                permissions: { type: array, minItems: 1, items: { type: string } }
      responses:
        "201": { description: Role created. }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /v1/users:
    get:
      operationId: listUsers
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Provisioned local and OIDC users. }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
    post:
      operationId: provisionUser
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role_ids]
              anyOf:
                - required: [oidc_issuer, oidc_subject]
                - required: [email, password]
              properties:
                oidc_issuer: { type: string, format: uri }
                oidc_subject: { type: string }
                email: { type: string, format: email }
                display_name: { type: string }
                password: { type: string, minLength: 12, description: Optional local self-hosted user password. }
                role_ids: { type: array, minItems: 1, items: { type: string, format: uuid } }
      responses:
        "201": { description: User provisioned. }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /v1/audit:
    get:
      operationId: listAuditEvents
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
      responses:
        "200":
          description: Most recent tenant/workspace audit events.
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    Error:
      description: Structured API error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Event:
      type: object
      additionalProperties: false
      required: [event_type, schema_version, idempotency_key, occurred_at, payload]
      anyOf:
        - required: [external_id]
        - required: [anonymous_id]
      properties:
        event_type:
          type: string
          examples: [profile.updated, consent.changed, product.viewed]
        schema_version:
          type: integer
          minimum: 1
        external_id:
          type: string
        anonymous_id:
          type: string
        idempotency_key:
          type: string
        occurred_at:
          type: string
          format: date-time
        source:
          type: string
        source_event_id:
          type: string
        correlation_id:
          type: string
        causation_id:
          type: string
        traceparent:
          type: string
        data_classification:
          enum: [public, internal, confidential, restricted]
        consent_context:
          type: object
          additionalProperties: true
        payload:
          type: object
          additionalProperties: true
    Profile:
      type: object
      required: [id, attributes, version, updated_at]
      properties:
        id:
          type: string
          format: uuid
        external_id:
          type: string
        anonymous_id:
          type: string
        attributes:
          type: object
          additionalProperties: true
        version:
          type: integer
        updated_at:
          type: string
          format: date-time
...
    AuthSession:
      type: object
      required: [access_token, token_type, expires_at]
      properties:
        access_token: { type: string }
        token_type: { const: Bearer }
        expires_at: { type: string, format: date-time }
...